+++ tomcat/src/main/java/com/techcourse/controller/LoginPageController.java
@@ -1,14 +1,15 @@
-package org.apache.coyote.http11.handler;
+package com.techcourse.controller;
 
 import java.io.IOException;
-import org.apache.coyote.http11.request.HttpProtocol;
+import org.apache.coyote.http11.HttpRequestHandler;
 import org.apache.coyote.http11.request.HttpRequest;
-import org.apache.coyote.http11.request.Method;
-import org.apache.coyote.http11.request.Uri;
+import org.apache.coyote.http11.request.line.HttpProtocol;
+import org.apache.coyote.http11.request.line.Method;
+import org.apache.coyote.http11.request.line.Uri;
 import org.apache.coyote.http11.response.HttpResponse;
 import org.apache.util.FileUtils;
 
-public class LoginPageHandler implements HttpRequestHandler {
+public class LoginPageController implements HttpRequestHandler {
 
     private static final String PAGE_RESOURCE_PATH = "/login.html";
     private static final Method SUPPORTING_METHOD = Method.GET;
