+++ tomcat/src/main/java/org/apache/coyote/http11/HttpRequestHandler.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.handler;
+package org.apache.coyote.http11;
 
 import java.io.IOException;
 import org.apache.coyote.http11.request.HttpRequest;
