+++ tomcat/src/main/java/org/apache/coyote/http11/request/line/HttpProtocol.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.request;
+package org.apache.coyote.http11.request.line;
 
 import com.techcourse.exception.UncheckedServletException;
 import java.util.Arrays;
