+++ tomcat/src/main/java/org/apache/coyote/http11/request/line/RequestLine.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.request;
+package org.apache.coyote.http11.request.line;
 
 public class RequestLine {
